requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy
albumentations>=1.1.0
Pillow>=8.0.0
tqdm
opencv-python
torchinfo
